Move out ReportCategory from constants
